IPv6 support for client/server
The server now has an option in the config file for 'dualmode(bool)'.
Setting this to true will enable dual mode sockets. The server must also
then bind to an IPv6 address (such as ::). The client will have to set
dualmode to true as well, which can be done in the ConnectedPlayer
class.